+++ src/music.f90
@@ -2,74 +2,76 @@
 !          algorithmic music and music theory
 ! License GPL-3.0-or-later
 ! Vincent Magnin
-! Last modifications: 2024-06-06
+! Last modifications: 2024-06-09
 
 module music
-    use, intrinsic :: iso_fortran_env, only: int8, int32, error_unit
-    use formidi, only : MIDI_delta_time, MIDI_Note, ON, OFF
-    ! Music theory elements common to the ForMIDI and ForSynth projects:
+    use, intrinsic :: iso_fortran_env, only: int8, error_unit
+    ! Music theory elements common to the ForMIDI and ForSynth projects.
+    ! We don't use "only:" to allow "use music" to include
+    ! the whole "music_common" module:
     use music_common
 
     implicit none
 
-    public :: write_chord, write_broken_chord, get_note_name
+    public :: get_note_name, get_MIDI_note
 
 contains
 
-    subroutine write_chord(channel, Note_MIDI, chord, velocity, duration)
-        ! Writes a chord, waits for its duration, and writes the OFF events
-        integer(int8), intent(in)  :: channel, Note_MIDI
-        integer, dimension(:), intent(in) :: chord
-        integer(int8),  intent(in) :: velocity
-        integer(int32), intent(in) :: duration
-        integer :: i
+    ! Returns the MIDI note number, from 12 (C0) to 127 (G9).
+    ! The note name is composed of two or three characters, 
+    ! for example "A4", "A#4", "Ab4", where the final character is 
+    ! the octave.
+    integer(int8) function get_MIDI_note(note)
+        character(*), intent(in) :: note
+        ! 0 <= octave <=9
+        integer(int8) :: octave
+        ! Gap relative to A4 (note 69) in semitones:
+        integer(int8) :: gap
+        ! ASCII code of the 0 character:
+        integer(int8), parameter :: zero = iachar('0')
 
-        do i = 1, size(chord)
-            call MIDI_delta_time(0)
-            call MIDI_Note(ON,  channel, Note_MIDI + int(chord(i), kind=int8), velocity)
-        end do
+        select case (note(1:1))
+            case ('C')
+                gap = -9
+            case ('D')
+                gap = -7
+            case ('E')
+                gap = -5
+            case ('F')
+                gap = -4
+            case ('G')
+                gap = -2
+            case ('A') 
+                gap = 0
+            case ('B')
+                gap = +2
+            case default
+                write(error_unit, *) "ERROR 4: unknown note name!"
+                error stop 4
+        end select
 
-        call MIDI_delta_time(duration)
+        ! Treating accidentals (sharp, flat) and computing the octave:
+        select case (note(2:2))
+            case ('b')
+                gap = gap - 1_int8
+                octave = iachar(note(3:3), int8) - zero
+            case ('#')
+                gap = gap + 1_int8
+                octave = iachar(note(3:3), int8) - zero
+            case default
+                octave = iachar(note(2:2), int8) - zero
+        end select
 
-        do i = 1, size(chord)
-            call MIDI_Note(OFF, channel, Note_MIDI + int(chord(i), kind=int8), 0_int8)
-            if (i < size(chord)) call MIDI_delta_time(0)
-        end do
-    end subroutine
+        if ((octave >= 0) .and. (octave <= 9)) then
+            gap = gap + (octave - 4_int8) * 12_int8
+        else
+            write(error_unit, *) "ERROR 5: octave out of bounds [0; 9]"
+            error stop 5
+        end if
 
-    ! Writes a broken chord using an array containing the intervals
-    ! (see the music_common module).
-    ! For the moment, each note has the same duration.
-    ! https://en.wikipedia.org/wiki/Arpeggio
-    subroutine write_broken_chord(channel, Note_MIDI, chord, velocity, duration)
-        integer(int8), intent(in)  :: channel, Note_MIDI
-        integer, dimension(:), intent(in) :: chord
-        integer(int8),  intent(in) :: velocity
-        integer(int32), intent(in) :: duration
-        integer(int32) :: dnote, residual
-        integer :: i
-
-        dnote = nint(real(duration) / size(chord))
-        ! The MIDI duration being an integer, the last note of the chord may
-        ! have a slightly different duration to keep the total duration exact:
-        residual = duration - dnote*(size(chord) - 1)
-
-        call MIDI_delta_time(0)
-        do i = 1, size(chord)
-            call MIDI_Note(ON,  channel, Note_MIDI + int(chord(i), kind=int8), velocity)
-            if (i < size(chord)) then 
-                call MIDI_delta_time(dnote)
-            else
-                call MIDI_delta_time(residual)
-            end if
-        end do
-
-        do i = 1, size(chord)
-            call MIDI_Note(OFF, channel, Note_MIDI + int(chord(i), kind=int8), 0_int8)
-            ! The delta time must always be placed before a note:
-            if (i < size(chord)) call MIDI_delta_time(0)
-        end do
-    end subroutine write_broken_chord
+        ! Computing and returning the MIDI note number (A4 is 69):
+        get_MIDI_note = 69_int8 + gap
+    end function get_MIDI_note
 
     ! Receives a MIDI note (for example 69),
     ! and returns the name of the note (for example A4).